Remove calls to PortableRemoteObject.narrow
The refactor was done by adding a javax.rmi.PortableRemoteObject class
to the project with a `narrow` method that simply returned the original
object.  The narrow method was then inlined.

Intellij will clean unused imports even if they do not relate to the
refactor, so there are a lot of .* imports that are also deleted.
I hope this is not an issue.

Signed-off-by: David Blevins <david.blevins@gmail.com>